refactor(abstract-utxo): move signTransaction to its own file
Issue: BTC-1450
