+++ modules/abstract-utxo/src/transaction/fixedScript/index.ts
@@ -1,4 +1,5 @@
 export { explainPsbt, explainLegacyTx, ChangeAddressInfo } from './explainTransaction';
 export { parseTransaction } from './parseTransaction';
-export { verifyTransaction } from './verifyTransaction';
 export { CustomChangeOptions } from './parseOutput';
+export { verifyTransaction } from './verifyTransaction';
+export { signTransaction } from './signTransaction';
